+++ man-roxygen/Hatchery_template.R
@@ -4,9 +4,11 @@
 #' [online documentation](https://docs.salmonmse.com/articles/equations.html#hatchery-production).
 #' @slot n_r Integer. Number of release strategies, sub-groups of fish with different survival schedules. Default is one.
 #' @slot n_yearling Vector length `n_r`. The target number of yearlings released by release strategy. No hatchery is modeled if `sum(n_yearling, n_subyearling) = 0`.
-#' are zero. Yearlings survival is density-independent after release. Default is zero.
+#' are zero. Default is zero. Yearlings are differentiated from sub-yearlings only in the hatchery survival parameter, e.g., lower egg-yearling survival than for egg-subyearling, so more brood is needed for yearling releases.
 #' @slot n_subyearling Vector length `n_r`. The target number of subyearlings released. No hatchery is modeled if `sum(n_yearling, n_subyearling) = 0` are zero.
-#' Subyearlings experience density-dependent survival in competition with natural origin young. Default is zero.
+#' Default is zero. Yearlings are differentiated from sub-yearlings only in the hatchery survival parameter, e.g., lower egg-yearling survival than for egg-subyearling, so more brood is needed for yearling releases.
+#' @slot yearling_DD Logical, whether freshwater yearling survival after release is density-dependent, in competition with natural-origin juveniles. Default is FALSE. Likely relevant only for stream types where there is substantial residence time before outmigration.
+#' @slot subyearling_DD Logical, whether freshwater subyearling survival after release is density-dependent, in competition with natural-origin juveniles. Default is FALSE. Likely relevant only for stream types where there is substantial residence time before outmigration.
 #' @slot s_prespawn Numeric. The survival of broodtake prior to egg production. `1 - s_prespawn` is the proportion of fish not used for hatchery purposes, e.g., mortality or other resesarch purposes.
 #' Used to back-calculate the broodtake from `n_yearling` and `n_subyearling`. Default is 1.
 #' @slot s_egg_smolt Numeric. The survival of eggs to the smolt life stage (for yearling release).